Remove OfficeQuoteServlet.java and modify DataServlet.java to complete its task.
